// src/dto/UpdateEmployeeDto.d.ts
export type UpdateEmployeeDto = {
    employeeId: number;
    name: string;
    email: string;
    phone: string;
    role: string;
    departmentId: number;
    designationId: number;
    status: string;
    joiningDate: Date;
    salary: number;
    gender: string;
    maritalStatus: string;
    address1: string;
    address2?: string;
    countryId: number;
    stateId: number;
    cityId: number;
    pincode: string;
    profileImage: string | null;
};
//# sourceMappingURL=UpdateEmployeeDto.d.ts.map